(svn r2597) Feature: [string system] Support cases.   - Cases are used to change a string, such as Coal Mine, depending on the surrounding context.   - Cases are defined like "STR_4802_COAL_MINE.ack  :Coala Mina"   - All cases need to be listed on the top of the file like this "##case ack"   - When using the string, type {STRING.ack} to choose the "ack" version of Coal mine.   - Also combined the strgen arrays into a struct, and fixed a bug with SetXY.

+++ strings.h
@@ -10,7 +10,7 @@ static inline char* InlineString(char* buf, uint16 string)
 }
 
 char *GetString(char *buffr, uint16 string);
-char *GetStringWithArgs(char *buffr, uint16 string, const int32 *argv);
+char *GetStringWithArgs(char *buffr, uint string, const int32 *argv);
 
 void InjectDParam(int amount);
 int32 GetParamInt32(void);
